pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selectools"
version = "0.8.0"
description = "Build AI agents that call your custom Python functions. Connect LLMs to tools, APIs, and databases with any provider (OpenAI, Anthropic, Gemini, Ollama)."
readme = "README.md"
requires-python = ">=3.9"
license = "LGPL-3.0-or-later"
authors = [{name = "John", email = "johnnichev@gmail.com"}]
dependencies = [
    "openai>=1.30.0,<2.0.0",
    "anthropic>=0.28.0,<1.0.0",
    "google-genai>=1.0.0",
    "numpy>=1.24.0,<3.0.0",
]
keywords = ["llm", "tools", "ai", "openai", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/johnnichev/selectools"
Repository = "https://github.com/johnnichev/selectools"
Issues = "https://github.com/johnnichev/selectools/issues"
Documentation = "https://github.com/johnnichev/selectools#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "build>=1.2.1",
    "twine>=5.0.0",
    "pre-commit>=3.6.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
]
rag = [
    "chromadb>=0.4.0",
    "pinecone-client>=3.0.0",
    "voyageai>=0.2.0",
    "cohere>=5.0.0",
    "pypdf>=4.0.0",
]

[project.scripts]
selectools = "selectools.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "selectools",
    "selectools.providers",
    "selectools.toolbox",
    "selectools.embeddings",
    "selectools.rag",
    "selectools.rag.stores",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "tests/",
    "examples/",
    "build/",
    "dist/",
]

[tool.bandit]
exclude_dirs = ["tests", "examples", "scripts"]
skips = ["B101", "B601", "B602"]  # Skip assert_used, shell, and subprocess checks
targets = ["src/"]
